Converted webMConverter functions into "pure functions"
Implemented basic unit testing

Added ffmpeg into package

Seperated user input and functionality

Added skeleton of cropVideo() and trimVideo()
